backend/requirements.txt
altair
amqp
annotated-types
anyio
appdirs
asgiref
attrs
Automat
beautifulsoup4
billiard
blinker
bs4
CacheControl
cachetools
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cloudscraper
constantly
coverage
crochet
cryptography
cssselect
defusedxml
distro
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework_simplejwt
dotenv
fake-useragent
filelock
firebase-admin
gevent
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-generativeai
google-resumable-media
googleapis-common-protos
greenlet
groq
grpcio
grpcio-status
gunicorn
h11
httpcore
httplib2
httpx
hyperlink
idna
incremental
iniconfig
itemadapter
itemloaders
Jinja2
jmespath
jsonschema
jsonschema-specifications
kombu
lxml
lxml_html_clean
MarkupSafe
msgpack
narwhals
numpy
outcome
packaging
pandas
parse
parsel
pillow
playwright
playwright-stealth
pluggy
prompt_toolkit
Protego
proto-plus
protobuf
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pydeck
PyDispatcher
pyee
Pygments
PyJWT
pyOpenSSL
pyparsing
pyppeteer
pyquery
PySocks
pytest
pytest-cov
pytest-django
python-dateutil
python-dotenv
pytz
queuelib
redis
referencing
requests
requests-file
requests-html
requests-toolbelt
rpds-py
rsa
Scrapy
selenium
service-identity
setuptools
six
smmap
sniffio
sortedcontainers
soupsieve
sqlparse
streamlit
tenacity
tldextract
toml
tornado
tqdm
trio
trio-websocket
Twisted
typing-inspection
typing_extensions
undetected-chromedriver
uritemplate
urllib3
vine
w3lib
watchdog
wcwidth
webdriver-manager
websocket-client
websockets
wrapt
wsproto
zope.event
zope.interface